+++ src/main/resources/l10n/JabRef_en.properties
@@ -2487,5 +2487,5 @@ Success\!\ Finished\ writing\ metadata.=Success! Finished writing metadata.
 
 Custom\ API\ key=Custom API key
 Check\ %0\ API\ Key\ Setting=Check %0 API Key Setting
-Ignore\ Unlinked\ Files=Ignore Unlinked Files
+Ignore\ File\ Extensions=Ignore File Extensions
 Include\ All\ Files=Include All Files
